refactor(docs): isolate runtime registry validation
